% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine-data.R
\docType{data}
\name{africa_mines}
\alias{africa_mines}
\title{Dataset containing the total number of mines and total number of mines by 24 commodities for each ADM2
district in Africa
#'}
\format{
A data frame with 88920 rows and 6 variables:
\describe{
\item{countrycode}{iso3c country code}
\item{mine_type}{type of mine}
\item{countryname}{country name}
\item{adm1}{ADM1 code}
\item{adm2}{ADM2 code}
\item{count}{# mines}
}
}
\source{
\url{https://data.world/datainspace/mines-in-africa}
}
\usage{
africa_mines
}
\description{
Dataset containing the total number of mines and total number of mines by 24 commodities for each ADM2
district in Africa
#'
}
\keyword{datasets}
