% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Africa_tilemap_package.R
\docType{package}
\name{AfricaCountryBins-package}
\alias{AfricaCountryBins-package}
\title{Create Country Bin maps of Africa!}
\description{
Create tile heatmaps as an alternative to choropleth maps to avoid distorted intuition
based on land mass size
This package is a modified version of Bob Rudis's statebins package, designed for use for data involving countries in
Africa. This grid attempts to preserve the general position of countries on the continent, while providing an equal
area by country.
}
\author{
Joshua de la Bruere (Joshua@delabj.com)

Bob Rudis (bob@rud.is)
}
