% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_africa_countrybins.R
\docType{data}
\name{geom_countrybins_africa}
\alias{geom_countrybins_africa}
\alias{GeomcountrybinsAfrica}
\title{A countrybins of Africa}
\format{An object of class \code{GeomcountrybinsAfrica} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.}
\usage{
geom_countrybins_africa(
  mapping = NULL,
  data = NULL,
  border_col = "white",
  border_size = 2,
  lbl_size = 3,
  dark_lbl = "black",
  light_lbl = "white",
  radius = grid::unit(6, "pt"),
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

GeomcountrybinsAfrica
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or
\code{aes_()}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{ggplot()}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{fortify()} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{border_col}{border color of the country squares, default "\code{white}"}

\item{border_size}{thickness of the square country borders}

\item{lbl_size}{font size (relative) of the label text}

\item{dark_lbl, light_lbl}{colrs to be uses when the label should be dark or light.
The function automagically computes when this should be.}

\item{radius}{the corner radius}

\item{...}{other arguments passed on to \code{layer()}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{borders()}.}
}
\description{
Based off the work of \href{https://github.com/hrbrmstr/statebins}{Bob Rudis's Statebins} this package aims to create a workable tile map of Africa for ggplot2.
Simply pass in a dataframe containing countries and values to measure, and this geom will create a simple map. This allows for easy faceting and allows a uniform legend across all plots.

Much like the origional statebins package, there are two critical \code{aes()} mappings: \cr
\itemize{
\item \code{country} (so the geom knows which column to map the country names/abbreviations to)
\item \code{fill} (the column you're mapping the filling with.)
}
}
\examples{
\dontrun{

library(AfricaCountryBins)
library(ggplot2)

data <- fake_data

ggplot(data, aes(country = country_code, fill = value))+
    geom_countrybins_africa()+
    theme_void()

}

}
\keyword{datasets}
